Stream output of compose commands and create script for distributing binaries
