--- couch-cli/.editorconfig
@@ -1,4 +1,0 @@
-[*.go]
-indent_style = space
-indent_size = 4
-tab_width = 4
