--- couch-cli/cmd/start.go
@@ -1,43 +1,0 @@
-/*
-Copyright © 2023 Michael Bui 293886@via.dk
-*/
-
-package cmd
-
-import (
-	"couch-cli/internal/service"
-	"log"
-
-	"github.com/spf13/cobra"
-)
-
-// startCmd represents the start command
-var startCmd = &cobra.Command{
-	Use:   "start",
-	Short: "Start a service",
-	Long: `Start a service. Additionally it will also search for any files named "service-account-key.json" 
-and .env in the service and apply the .env file
-`,
-	Args: cobra.ExactArgs(1),
-	RunE: func(cmd *cobra.Command, args []string) error {
-		err := service.StartService(args[0])
-		if err != nil {
-			log.Fatal(err)
-		}
-		return nil
-	},
-}
-
-func init() {
-	rootCmd.AddCommand(startCmd)
-
-	// Here you will define your flags and configuration settings.
-
-	// Cobra supports Persistent Flags which will work for this command
-	// and all subcommands, e.g.:
-	// startCmd.PersistentFlags().String("foo", "", "A help for foo")
-
-	// Cobra supports local flags which will only run when this command
-	// is called directly, e.g.:
-	// startCmd.Flags().BoolP("toggle", "t", false, "Help message for toggle")
-}
