--- couch-cli/go.mod
@@ -1,12 +1,0 @@
-module couch-cli
-
-go 1.18
-
-require gopkg.in/yaml.v3 v3.0.1
-
-require (
-	github.com/go-yaml/yaml v2.1.0+incompatible // indirect
-	github.com/inconshreveable/mousetrap v1.1.0 // indirect
-	github.com/spf13/cobra v1.7.0 // indirect
-	github.com/spf13/pflag v1.0.5 // indirect
-)
