+++ devtools/couch-cli/.editorconfig
@@ -1,0 +1,4 @@
+[*.go]
+indent_style = space
+indent_size = 4
+tab_width = 4
