+++ devtools/couch-cli/cmd/new.go
@@ -1,0 +1,30 @@
+/*
+Copyright © 2023 Michael Bui 293886@via.dk
+*/
+
+package cmd
+
+import (
+	"github.com/spf13/cobra"
+)
+
+// newCmd represents the new command
+var newCmd = &cobra.Command{
+	Use:   "new",
+	Short: "Create new <SUBCOMMAND>",
+	Long:  `Create new <SUBCOMMAND>`,
+}
+
+func init() {
+	rootCmd.AddCommand(newCmd)
+
+	// Here you will define your flags and configuration settings.
+
+	// Cobra supports Persistent Flags which will work for this command
+	// and all subcommands, e.g.:
+	// newCmd.PersistentFlags().String("foo", "", "A help for foo")
+
+	// Cobra supports local flags which will only run when this command
+	// is called directly, e.g.:
+	// newCmd.Flags().BoolP("toggle", "t", false, "Help message for toggle")
+}
